+++ application/cloudevent-converter/src/main/java/org/occurrent/application/converter/CloudEventConverter.java
@@ -18,6 +18,8 @@ package org.occurrent.application.converter;
 
 import io.cloudevents.CloudEvent;
 
+import java.util.stream.Stream;
+
 /**
  * A cloud event converter interface that is used by Occurrent application services
  * to convert to and from domain events.
@@ -25,6 +27,19 @@ import io.cloudevents.CloudEvent;
  * @param <T> The type of your domain event
  */
 public interface CloudEventConverter<T> {
+
+    /**
+     * Convert a stream of domain events into a stream of cloud events.
+     * The reason for implementing this method is to allow adding things such as correlation id
+     * that should be the same for all events in a stream.
+     *
+     * @param events The domain events to convert to cloud events
+     * @return A stram of cloud events.
+     */
+    default Stream<CloudEvent> toCloudEvents(Stream<T> events) {
+        Stream<T> stream = events == null ? Stream.empty() : events;
+        return stream.map(this::toCloudEvent);
+    }
 
     /**
      * Convert a domain event into a cloud event
@@ -41,4 +56,16 @@ public interface CloudEventConverter<T> {
      * @return The domain event instance, converted from the cloud event.
      */
     T toDomainEvent(CloudEvent cloudEvent);
+
+    /**
+     * Convert a stream of cloud events into a stream of domain events.
+     * Typically, you only need to implement {@link #toDomainEvent(CloudEvent)}.
+     *
+     * @param events The cloud events to convert to domain events
+     * @return A stram of cloud events.
+     */
+    default Stream<T> toDomainEvents(Stream<CloudEvent> events) {
+        Stream<CloudEvent> stream = events == null ? Stream.empty() : events;
+        return stream.map(this::toDomainEvent);
+    }
 }