+++ application/service/blocking/src/main/java/org/occurrent/application/service/blocking/generic/GenericApplicationService.java
@@ -84,7 +84,7 @@ public class GenericApplicationService<T> implements ApplicationService<T> {
             EventStream<CloudEvent> eventStream = eventStore.read(streamId);
 
             // Convert the cloud events into domain events
-            Stream<T> eventsInStream = eventStream.events().map(cloudEventConverter::toDomainEvent);
+            Stream<T> eventsInStream = cloudEventConverter.toDomainEvents(eventStream.events());
 
             // Call a pure function from the domain model which returns a Stream of events
             Stream<T> newDomainEvents = emptyStreamIfNull(functionThatCallsDomainModel.apply(eventsInStream));
@@ -94,7 +94,7 @@ public class GenericApplicationService<T> implements ApplicationService<T> {
             final List<T> newEventsAsList = sideEffect == null ? null : newDomainEvents.collect(Collectors.toList());
 
             // Convert to cloud events and write the new events to the event store
-            Stream<CloudEvent> newEvents = (sideEffect == null ? newDomainEvents : newEventsAsList.stream()).map(cloudEventConverter::toCloudEvent);
+            Stream<CloudEvent> newEvents = cloudEventConverter.toCloudEvents(sideEffect == null ? newDomainEvents : newEventsAsList.stream());
             WriteResult writeResult = eventStore.write(streamId, eventStream.version(), newEvents);
             return new Tuple<>(writeResult, newEventsAsList);
         });